Final-Project/Project-3-Parser/requirements.txt
# Requirements for Final Project 3: Parser
# pip install -r requirements.txt

langchain
langchain-openai
python-dotenv
pydantic
ipykernel
